Backend wiring for minisearch app
This commit stores the basic wiring for the minisearch application.
It should be combined with the minisearch frontend Ember application.
Although this commit contains some functionality it is by no means to
be considered stable yet.  There are almost no version numbers in the
docker-compose.yml, the configuration itself is in flux, and there are
some known issues with the resulting functionality.